Rename and some clean up
