+++ eng/versioning/version_client.txt
@@ -89,7 +89,6 @@ com.azure:azure-core-experimental;1.0.0-beta.52;1.0.0-beta.53
 com.azure:azure-core-http-jdk-httpclient;1.0.0-beta.15;1.0.0-beta.16
 com.azure:azure-core-http-netty;1.15.3;1.16.0-beta.1
 com.azure:azure-core-http-okhttp;1.12.2;1.13.0-beta.1
-com.azure:azure-core-http-test-common;1.0.0-beta.1;1.0.0-beta.1
 com.azure:azure-core-http-vertx;1.0.0-beta.20;1.0.0-beta.21
 com.azure:azure-core-management;1.15.2;1.16.0-beta.1
 com.azure:azure-core-metrics-opentelemetry;1.0.0-beta.21;1.0.0-beta.22
