+++ sdk/core/azure-core-http-jdk-httpclient/src/test/java/com/azure/core/http/jdk/httpclient/JdkHttpClientLocalTestServer.java
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 package com.azure.core.http.jdk.httpclient;
 
-import com.azure.core.test.shared.LocalTestServer;
+import com.azure.core.validation.http.LocalTestServer;
 import org.eclipse.jetty.util.Callback;
 
 import javax.servlet.ServletException;
