+++ sdk/core/azure-core-http-okhttp/src/test/java/com/azure/core/http/okhttp/OkHttpClientLocalTestServer.java
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 package com.azure.core.http.okhttp;
 
-import com.azure.core.test.shared.LocalTestServer;
+import com.azure.core.validation.http.LocalTestServer;
 import org.eclipse.jetty.util.Callback;
 
 import javax.servlet.ServletException;
