+++ sdk/core/azure-core-http-vertx/src/test/java/com/azure/core/http/vertx/VertxAsyncHttpClientAsyncHttpClientTests.java
@@ -4,9 +4,9 @@
 package com.azure.core.http.vertx;
 
 import com.azure.core.http.HttpClient;
-import com.azure.core.test.shared.HttpClientTests;
-import com.azure.core.test.shared.HttpClientTestsServer;
-import com.azure.core.test.shared.LocalTestServer;
+import com.azure.core.validation.http.HttpClientTests;
+import com.azure.core.validation.http.HttpClientTestsServer;
+import com.azure.core.validation.http.LocalTestServer;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.parallel.Execution;
