+++ sdk/core/azure-core/src/test/java/com/azure/core/validation/http/SyncAsyncTest.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.test.shared;
+package com.azure.core.validation.http;
 
 import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
