+++ sdk/core/azure-core/src/test/java/com/azure/core/validation/http/models/HttpBinJson.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.core.test.shared.models;
+package com.azure.core.validation.http.models;
 
 import com.azure.json.JsonReader;
 import com.azure.json.JsonSerializable;
