+++ sdk/core/azure-core/src/test/resources/keystore.jks
